Refactor components into separate files
I think it's going to get a little messy to keep all of the components
in one file, so I've refactored them into separate files in a components
directory. I'm thinking it might be a good idea to build charts as
helpers in a separate directory that can be constructed and applied
using d3.selection.call, similarly to how D3 implements axes.
