+++ services/ui/src/js/main.js
@@ -1,5 +1,7 @@
 import styles from "../styles/_index.scss"; // eslint-disable-line
-import * as diag from "./diagnostics";
 
-customElements.define("ug-diag-scalarvariable", diag.ScalarVariableDiag);
-customElements.define("ug-diag-scalarloop", diag.ScalarLoopDiag);
+import ScalarVariableDiag from "./components/ScalarVariableDiag";
+import ScalarLoopDiag from "./components/ScalarLoopDiag";
+
+customElements.define("ug-diag-scalarvariable", ScalarVariableDiag);
+customElements.define("ug-diag-scalarloop", ScalarLoopDiag);
